Move common code to homeassistant/util/redact.py
